modified:   app/Customs/Services/EmailVerificationService.php 	modified:   app/Http/Controllers/AuthController.php 	modified:   app/Http/Controllers/CartController.php 	modified:   app/Http/Controllers/CategoryController.php 	modified:   app/Http/Controllers/LocationController.php 	modified:   app/Http/Controllers/ProductController.php 	deleted:    app/Http/Requests/LoginRequest.php 	deleted:    app/Http/Requests/RegistrationRequest.php 	deleted:    app/Http/Requests/ResendVerificationEmailCodeRequest.php 	deleted:    app/Http/Requests/VerifyEmailRequest.php 	modified:   database/migrations/2024_05_09_121030_create_cart_items_table.php 	modified:   routes/api.php

+++ app/Http/Controllers/LocationController.php
@@ -7,8 +7,25 @@ use Illuminate\Http\Request;
 use Auth;
 use Validator;
 use Exception;
+
 class LocationController extends Controller
 {
+    public function get_user_locations()
+    {
+        $user_id = auth()->id();
+
+        $locations = Location::where('user_id', $user_id)->get();
+
+        if ($locations->isNotEmpty()) {
+            return response()->json(['locations'=>$locations], 200);
+        } else {
+            return response()->json([
+                'status' => 'success',
+                'message' => "user have no locations"
+            ]);
+        }
+
+    }
 
     /**
      * Store a newly created resource in storage.
@@ -16,7 +33,7 @@ class LocationController extends Controller
     public function store(Request $request)
     {
         try {
-            $validator=Validator::make($request->all(), [
+            $validator = Validator::make($request->all(), [
                 'city' => 'required',
                 'street' => 'required',
                 'building' => 'required',
@@ -27,8 +44,10 @@ class LocationController extends Controller
                 'building' => $request->building,
                 'user_id' => Auth::id(),
             ]);
-            return response()->json('Location added', 201);
-
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Location added'
+            ], 201);
         } catch (Exception $e) {
             $data = [$e, $validator->errors()];
 
@@ -42,24 +61,35 @@ class LocationController extends Controller
     public function update(Request $request, $id)
     {
         try {
-        $validator=Validator::make($request->all(), [
-            'city' => 'required',
-            'street' => 'required',
-            'building' => 'required',
-        ]);
+            $validator = Validator::make($request->all(), [
+                'city' => 'required',
+                'street' => 'required',
+                'building' => 'required',
+            ]);
 
-        $location = Location::find($id);
+            $location = Location::find($id);
+            if ($location->user_id != Auth::id()) {
+                return response()->json([
+                    'status' => 'failed',
+                    'message' => 'you dont have permission to edit this location'
+                ], 200);
+            }
+            if ($location) {
 
-        if ($location) {
+                $location->city = $request->city;
+                $location->street = $request->street;
+                $location->building = $request->building;
+                $location->save();
 
-            $location->city = $request->city;
-            $location->street = $request->street;
-            $location->building = $request->building;
-            $location->save();
-
-            return response()->json('location updated', 200);
-        } else
-            return response()->json('location not found');
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'location updated'
+                ], 200);
+            } else
+                return response()->json([
+                    'status' => 'failed',
+                    'message' => 'location not found'
+                ], 200);
         } catch (Exception $e) {
             $data = [$e, $validator->errors()];
 
@@ -73,11 +103,22 @@ class LocationController extends Controller
     public function destroy($id)
     {
         $location = Location::find($id);
+        if ($location->user_id != Auth::id()) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => 'you dont have permission to delete this location'
+            ], 200);
+        }
         if ($location) {
             $location->delete();
-            return response()->json('location deleted', 200);
+            return response()->json([
+                'status' => 'success',
+                'message' => 'location deleted'
+            ], 200);
         } else
-            return response()->json('location not found');
-
+            return response()->json([
+                'status' => 'failed',
+                'message' => 'location not found'
+            ], 200);
     }
 }
